+++ src/hyperactive/opt/_adapters/__init__.py
@@ -1,2 +1,7 @@
 """Adapters for individual packages."""
 # copyright: hyperactive developers, MIT License (see LICENSE file)
+
+from ._base_optuna_adapter import _BaseOptunaAdapter
+from ._gfo import _BaseGFOadapter
+
+__all__ = ["_BaseOptunaAdapter", "_BaseGFOadapter"]
