+++ src/hyperactive/opt/_adapters/_base_optuna_adapter.py
