+++ src/hyperactive/opt/optuna/_cmaes_sampler.py
@@ -1,7 +1,7 @@
 """CMA-ES (Covariance Matrix Adaptation Evolution Strategy) sampler optimizer."""
 # copyright: hyperactive developers, MIT License (see LICENSE file)
 
-from ._base_optuna_adapter import _BaseOptunaAdapter
+from .._adapters._base_optuna_adapter import _BaseOptunaAdapter
 
 
 class CmaEsSampler(_BaseOptunaAdapter):
