+++ src/hyperactive/opt/optuna/_gp_sampler.py
@@ -1,7 +1,7 @@
 """Gaussian Process sampler optimizer."""
 # copyright: hyperactive developers, MIT License (see LICENSE file)
 
-from ._base_optuna_adapter import _BaseOptunaAdapter
+from .._adapters._base_optuna_adapter import _BaseOptunaAdapter
 
 
 class GPSampler(_BaseOptunaAdapter):
