+++ src/hyperactive/opt/optuna/_grid_sampler.py
@@ -1,7 +1,7 @@
 """Grid sampler optimizer."""
 # copyright: hyperactive developers, MIT License (see LICENSE file)
 
-from ._base_optuna_adapter import _BaseOptunaAdapter
+from .._adapters._base_optuna_adapter import _BaseOptunaAdapter
 
 
 class GridSampler(_BaseOptunaAdapter):
