+++ src/hyperactive/opt/optuna/_nsga_ii_sampler.py
@@ -1,7 +1,7 @@
 """NSGA-II multi-objective sampler optimizer."""
 # copyright: hyperactive developers, MIT License (see LICENSE file)
 
-from ._base_optuna_adapter import _BaseOptunaAdapter
+from .._adapters._base_optuna_adapter import _BaseOptunaAdapter
 
 
 class NSGAIISampler(_BaseOptunaAdapter):
