+++ src/hyperactive/opt/optuna/_qmc_sampler.py
@@ -1,7 +1,7 @@
 """Quasi-Monte Carlo sampler optimizer."""
 # copyright: hyperactive developers, MIT License (see LICENSE file)
 
-from ._base_optuna_adapter import _BaseOptunaAdapter
+from .._adapters._base_optuna_adapter import _BaseOptunaAdapter
 
 
 class QMCSampler(_BaseOptunaAdapter):
