+++ src/hyperactive/opt/optuna/_random_sampler.py
@@ -1,7 +1,7 @@
 """Random sampler optimizer."""
 # copyright: hyperactive developers, MIT License (see LICENSE file)
 
-from ._base_optuna_adapter import _BaseOptunaAdapter
+from .._adapters._base_optuna_adapter import _BaseOptunaAdapter
 
 
 class RandomSampler(_BaseOptunaAdapter):
