+++ src/hyperactive/opt/optuna/_tpe_sampler.py
@@ -1,7 +1,7 @@
 """TPE (Tree-structured Parzen Estimator) sampler optimizer."""
 # copyright: hyperactive developers, MIT License (see LICENSE file)
 
-from ._base_optuna_adapter import _BaseOptunaAdapter
+from .._adapters._base_optuna_adapter import _BaseOptunaAdapter
 
 
 class TPESampler(_BaseOptunaAdapter):
